chore(tests): gather river fixtures in river dir
