+++ tests/river/conftest.py
@@ -1,10 +1,18 @@
 import inspect
+import json
+import time
+from pathlib import Path
 
 import pytest
+from confluent_kafka import admin
 from factory import Factory
 from pytest_factoryboy import register
 
+from django.conf import settings
+
 from . import factories
+
+DATA_FIXTURES_DIR = Path(__file__).resolve().parent / "fixtures"
 
 register(factories.BatchFactory)
 register(factories.ErrorFactory)
@@ -22,3 +30,41 @@ def reset_factories_sequences():
     """Reset all sequences for predictable values."""
     for factory in get_factories():
         factory.reset_sequence()
+
+
+@pytest.fixture(autouse=True)
+def clear_topics(request):
+    """Clear topics after each test marked with `kafka`."""
+
+    def _clear_topics():
+        admin_client = admin.AdminClient({"bootstrap.servers": settings.KAFKA_BOOTSTRAP_SERVERS})
+
+        def list_active_topics():
+            return [topic for topic in admin_client.list_topics().topics if topic != "__consumer_offsets"]
+
+        active_topics = list_active_topics()
+
+        if len(active_topics) > 0:
+            admin_client.delete_topics(active_topics, operation_timeout=1)
+
+            # Wait for the topics to be deleted. Note: the futures returned by
+            # `delete_topics` are unreliable to determine whether the topic deletion
+            # has been propagated.
+            while len(active_topics) > 0:
+                time.sleep(0.1)
+                active_topics = list_active_topics()
+
+    if request.node.get_closest_marker("kafka"):
+        request.addfinalizer(_clear_topics)
+
+
+@pytest.fixture
+def patient_mapping():
+    with (DATA_FIXTURES_DIR / "patient_mapping.json").open() as f:
+        return json.load(f)
+
+
+@pytest.fixture
+def users_to_patients_mapping():
+    with (DATA_FIXTURES_DIR / "users_to_patients_mapping.json").open() as f:
+        return json.load(f)
